26502 add app-name to logging metadata
