+++ api/namex/__init__.py
@@ -19,7 +19,6 @@ jwt = JwtManager()
 
 from flask_cors import CORS
 from flask_migrate import Migrate
-from structured_logging import StructuredLogging
 
 from namex.services.cache import cache
 from namex.services.lookup import nr_filing_actions
@@ -29,7 +28,7 @@ from namex import models
 from namex.models import db, ma
 from namex.resources import api
 from namex.utils.run_version import get_run_version
-from namex.services import flags
+from namex.services import flags, logging_config
 
 
 run_version = get_run_version()
@@ -41,22 +40,7 @@ def create_app(run_mode=os.getenv('FLASK_ENV', 'production')):  # noqa: B008
     CORS(app)
     app.config.from_object(config.CONFIGURATION[run_mode])
 
-    # Configure Structured Logging
-    structured_logger = StructuredLogging()
-    structured_logger.init_app(app)
-    app.logger = structured_logger.get_logger()
-
-    # Add support for request-specific log fields like 'app_name'
-    structlog.configure(
-        processors=[structlog.contextvars.merge_contextvars] + structlog.get_config()['processors']
-    )
-
-    # Bind app-name from request headers into the logging context
-    @app.before_request
-    def bind_app_name():
-        structlog.contextvars.clear_contextvars()
-        structlog.contextvars.bind_contextvars(app_name=request.headers.get('App-Name', 'unknown'))
-
+    logging_config.configure_logging(app)
     flags.init_app(app)
     queue.init_app(app)
